# agri_uav_disease_system/backend/docs/integration_package/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "三下乡无人机水稻病虫害识别系统",
    "version": "mvp-1"
  },
  "paths": {
...
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "batch-tasks"
        ],
        "summary": "Get Batch Task",
        "operationId": "get_batch_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchTaskStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/weather/observations": {
      "post": {
        "tags": [
          "weather"
        ],
        "summary": "Create Weather Observation",
        "operationId": "create_weather_observation_api_weather_observations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherObservationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherObservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "List Weather Observations",
        "operationId": "list_weather_observations_api_weather_observations_get",
        "parameters": [
          {
            "name": "plot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plot Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherObservationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchTaskStatus": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": "batch_image_detection"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_images": {
            "type": "integer",
            "title": "Total Images"
          },
          "processed_images": {
            "type": "integer",
            "title": "Processed Images"
          },
          "failed_images": {
            "type": "integer",
            "title": "Failed Images"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "record_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Record Ids"
          },
          "failed_items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Failed Items"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status",
          "total_images",
          "processed_images",
          "failed_images",
          "progress",
          "record_ids",
          "failed_items",
          "created_at",
          "updated_at"
        ],
        "title": "BatchTaskStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeatherObservation": {
        "properties": {
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "region_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Name"
          },
          "observed_date": {
            "type": "string",
            "title": "Observed Date"
          },
          "temperature_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Max"
          },
          "temperature_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Min"
          },
          "humidity_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity Avg"
          },
          "rainfall_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rainfall Mm"
          },
          "wind_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed"
          },
          "sunshine_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sunshine Hours"
          },
          "weather_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Text"
          },
          "data_source": {
            "type": "string",
            "title": "Data Source",
            "default": "manual"
          },
          "weather_id": {
            "type": "string",
            "title": "Weather Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "observed_date",
          "weather_id",
          "created_at"
        ],
        "title": "WeatherObservation"
      },
      "WeatherObservationCreate": {
        "properties": {
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "region_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Name"
          },
          "observed_date": {
            "type": "string",
            "title": "Observed Date"
          },
          "temperature_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Max"
          },
          "temperature_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Min"
          },
          "humidity_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity Avg"
          },
          "rainfall_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rainfall Mm"
          },
          "wind_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed"
          },
          "sunshine_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sunshine Hours"
          },
          "weather_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Text"
          },
          "data_source": {
            "type": "string",
            "title": "Data Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "observed_date"
        ],
        "title": "WeatherObservationCreate"
      },
      "WeatherObservationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WeatherObservation"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "WeatherObservationListResponse"
      }
    }
  }
}